pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.envs.default]
installer = "uv"

[tool.hatch.version]
source = "vcs"
fallback-version = "0.0.0.dev0"
tag-pattern = '''^[a-z]+\/(?:[\w-]+-)?(?P<version>[vV]?\d+(?:\.\d+){0,2}[^\+]*)(?:\+.*)?$'''

[tool.hatch.build.hooks.vcs]
version-file = "src/_version.py"

[tool.ruff]
line-length = 88
extend-exclude = ["src/_version.py", "dev-pyproject/*"]
respect-gitignore = true

[tool.ruff.lint]
select = [
  "E4",
  "E7",
  "E9",
  "F",
  "B",
  "Q",
  "I",
  "ARG",
  "ANN",
  "D",
  "PL",
  "S",
  "RUF",
  "UP",
  "TC",
  "SIM",
]
ignore = ["PLR2004", "PLR0913", "ANN401", "S101", "S311"]

[tool.ruff.lint.pylint]
max-args = 10

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.pyright]
analyzeUnannotatedFunctions = true
reportMissingParameterType = "error"
reportUnusedImport = "error"
reportUnusedClass = "error"
reportUnusedFunction = "error"
reportUnusedVariable = "error"
extraPaths = []

[tool.pytest.ini_options]
pythonpath = "./"

[tool.typos.files]
extend-exclude = ["**/*.env"]

[tool.uv.sources]
skypilot = { git = "https://github.com/AlonKellner/skypilot-beta.git", rev = "custom-new" }

[project]
dynamic = ["version"]
name = "finetune-setup-example"
requires-python = ">=3.10.0"
authors = [{ name = "Alon Kellner", email = "me@alonkellner.com" }]
maintainers = [{ name = "Alon Kellner", email = "me@alonkellner.com" }]
description = "An experimental finetuning setup example"
readme = "README.md"

keywords = []

classifiers = [
  "Development Status :: 3 - Alpha",

  "Intended Audience :: Science/Research",
  "Topic :: Multimedia :: Sound/Audio :: Analysis",

  "License :: OSI Approved :: MIT License",

  "Programming Language :: Python :: 3",
]

dependencies = ["inquirer"]

[dependency-groups]
dev_tools = ["pytest", "pyright"]
jobs = ["gitpython", "pyyaml", "skypilot[kubernetes,runpod]"]
gpu = ["flash-attn"]
mms_blog_post = [
  "accelerate",
  "boto3",
  "comet-ml",
  "datasets[audio]",
  "evaluate",
  "huggingface_hub",
  "jiwer",
  "llvmlite>0.36",
  "numba>0.47",
  "numpy",
  "pandas",
  "polars",
  "safetensors",
  "torch",
  "torchaudio",
  "transformers",
  "types-boto3-lite[s3]",
  "uroman",
  "wandb",
]
mms_blog_post_gpu = [
  { include-group = "mms_blog_post" },
  { include-group = "gpu" },
]
dev = [
  { include-group = "dev_tools" },
  { include-group = "mms_blog_post" },
  { include-group = "jobs" },
]
dev_gpu = [{ include-group = "dev" }, { include-group = "gpu" }]

[project.urls]
Homepage = "https://alonkellner.com"
Repository = "https://github.com/AlonKellner/finetune-setup-example.git"
